memory storage migrated, package.json updated to work with typescript
