+++ package.json
@@ -7,10 +7,13 @@
   },
   "type": "module",
   "description": "nodejs-mitso",
-  "main": "src/server.js",
+  "main": "build/server.js",
   "scripts": {
-    "start": "nodemon src/server.js",
-    "lint": "eslint --ext .js,.ts --fix src/"
+    "start": "node build/server.js",
+    "build": "tsc",
+    "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/server.ts",
+    "lint": "eslint 'src/**/*.ts' --fix",
+    "lint-staged": "lint-staged"
   },
   "keywords": [
     "rest",
@@ -35,8 +38,8 @@
     }
   },
   "lint-staged": {
-    "src/**/*.js": [
-      "npm run lint"
+    "src/**/*.ts": [
+      "eslint --fix"
     ]
   },
   "dependencies": {
